Refactoring fire,shield overlay gathered to NoOverlay.
